Blueprint Foldering, New Level Select menu with Tab
Press tab to open menu and click buttons to select a level or quite.
Problem if you don't disable game input, your player still moves around
and shoots. If you disable game input, however, you cant seem to use the
Menu InputAction (Tab) to escape the menu once you're in it.

Could just add a Resume button, but we'll need to have a better grasp of
UI's than that.

Also, how to simply pause the game?